First rewrite of lib/ in js, focus on structure
This library can be required and it will return similarly to the gem: an
instance is not required. However, most of the lib/percy/ code is not
implemented; only the main structure of the modules and how they're made
and stored is replicated.

The code translation has been as 1:1 as possible. It went as far as
using Class getters/setters to create their properties when they are
accessed rather than in the constructor: the end result is the same in
that there's no difference in value between the first access of the
property and when the instance was created. The only difference is time
of execution, which won't matter here because nothing is time sensitive.

**Note: This is definitely an anti-pattern in js.**

An advantage of the getter/setter usage here is that we can disallow
setting a property by simply not including a setter, so a property
cannot be changed.